clean up headers
Introduce include/ dir for public headers, and clean up include guards etc.
