+++ include/pci_caps/common.h
@@ -1,9 +1,7 @@
 /*
- * Copyright (c) 2019 Nutanix Inc. All rights reserved.
+ * Copyright (c) 2020 Nutanix Inc. All rights reserved.
  *
  * Authors: Thanos Makatos <thanos@nutanix.com>
- *          Swapnil Ingle <swapnil.ingle@nutanix.com>
- *          Felipe Franciosi <felipe@nutanix.com>
  *
  *  Redistribution and use in source and binary forms, with or without
  *  modification, are permitted provided that the following conditions are met:
@@ -30,43 +28,27 @@
  *
  */
 
-#ifndef LM_PCI_CAP_PM_H
-#define LM_PCI_CAP_PM_H
+#ifndef LIB_MUSER_PCI_CAPS_COMMON_H
+#define LIB_MUSER_PCI_CAPS_COMMON_H
 
-#include "common.h"
+#include <stddef.h>
 
-struct pc {
-    unsigned int vs:3;
-    unsigned int pmec:1;
-    unsigned int res:1;
-    unsigned int dsi:1;
-    unsigned int auxc:3;
-    unsigned int d1s:1;
-    unsigned int d2s:1;
-    unsigned int psup:5;
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct cap_hdr {
+    uint8_t id;
+    uint8_t next;
 } __attribute__((packed));
-_Static_assert(sizeof(struct pc) == 0x2, "bad PC size");
+_Static_assert(sizeof(struct cap_hdr) == 0x2, "bad PCI capability header size");
+_Static_assert(offsetof(struct cap_hdr, id) == PCI_CAP_LIST_ID, "bad offset");
+_Static_assert(offsetof(struct cap_hdr, next) == PCI_CAP_LIST_NEXT, "bad offset");
 
-struct pmcs {
-    unsigned int ps:2;
-    unsigned int res1:1;
-    unsigned int nsfrst:1;
-    unsigned int res2:4;
-    unsigned int pmee:1;
-    unsigned int dse:4;
-    unsigned int dsc:2;
-    unsigned int pmes:1;
-} __attribute__((packed));
-_Static_assert(sizeof(struct pc) == 0x2, "bad PMCS size");
+#ifdef __cplusplus
+}
+#endif
 
-struct pmcap {
-    struct cap_hdr hdr;
-    struct pc pc;
-    struct pmcs pmcs;
-} __attribute__((packed)) __attribute__ ((aligned(8))); /* FIXME why does it need to be aligned? */
-_Static_assert(sizeof(struct pmcap) == PCI_PM_SIZEOF, "bad PC size");
-_Static_assert(offsetof(struct pmcap, hdr) == 0, "bad offset");
-
-#endif /* LM_CAP_PM_H */
+#endif /* LIB_MUSER_PCI_CAPS_COMMON_H */
 
 /* ex: set tabstop=4 shiftwidth=4 softtabstop=4 expandtab: */
