+++ include/pci_caps/pm.h
@@ -30,40 +30,51 @@
  *
  */
 
-#ifndef LM_PCI_CAP_MSI_H
-#define LM_PCI_CAP_MSI_H
+#ifndef LIB_MUSER_PCI_CAPS_PM_H
+#define LIB_MUSER_PCI_CAPS_PM_H
 
 #include "common.h"
 
-struct mc {
-    unsigned int msie:1;
-    unsigned int mmc:3;
-    unsigned int mme:3;
-    unsigned int c64:1;
-    unsigned int pvm:1;
-    unsigned int res1:7;
-} __attribute__ ((packed));
-_Static_assert(sizeof(struct mc) == 0x2, "bad MC size");
+#ifdef __cplusplus
+extern "C" {
+#endif
 
-struct ma {
-    unsigned int res1:2;
-    unsigned int addr:30;
-} __attribute__ ((packed));
-_Static_assert(sizeof(struct ma) == 0x4, "bad MA size");
+struct pc {
+    unsigned int vs:3;
+    unsigned int pmec:1;
+    unsigned int res:1;
+    unsigned int dsi:1;
+    unsigned int auxc:3;
+    unsigned int d1s:1;
+    unsigned int d2s:1;
+    unsigned int psup:5;
+} __attribute__((packed));
+_Static_assert(sizeof(struct pc) == 0x2, "bad PC size");
 
-struct msicap {
+struct pmcs {
+    unsigned int ps:2;
+    unsigned int res1:1;
+    unsigned int nsfrst:1;
+    unsigned int res2:4;
+    unsigned int pmee:1;
+    unsigned int dse:4;
+    unsigned int dsc:2;
+    unsigned int pmes:1;
+} __attribute__((packed));
+_Static_assert(sizeof(struct pc) == 0x2, "bad PMCS size");
+
+struct pmcap {
     struct cap_hdr hdr;
-    struct mc mc;
-    struct ma ma;
-    uint32_t mua;
-    uint16_t md;
-    uint16_t padding;
-    uint32_t mmask;
-    uint32_t mpend;
-}  __attribute__ ((packed));
-_Static_assert(sizeof(struct msicap) == 0x18, "bad MSICAP size");
-_Static_assert(offsetof(struct msicap, hdr) == 0, "bad offset");
+    struct pc pc;
+    struct pmcs pmcs;
+} __attribute__((packed)) __attribute__ ((aligned(8))); /* FIXME why does it need to be aligned? */
+_Static_assert(sizeof(struct pmcap) == PCI_PM_SIZEOF, "bad PC size");
+_Static_assert(offsetof(struct pmcap, hdr) == 0, "bad offset");
 
-#endif /* LM_CAP_MSI_H */
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* LM_CAP_PM_H */
 
 /* ex: set tabstop=4 shiftwidth=4 softtabstop=4 expandtab: */
